requirements.txt
praw
kafka-python
scikit-learn
python-dotenv